separated crawling and parsing
